fix(bridge): rename bridge/types.py → bridge/bt_types.py to unshadow stdlib `types`
Real-world bug discovered after rc.3 deploy: every Sendspin daemon
subprocess crashed at start with:

    ImportError: cannot import name 'GenericAlias' from partially
    initialized module 'types' (most likely due to a circular import)
    (/usr/local/lib/python3.12/site-packages/sendspin_bridge/bridge/types.py)

User symptom: BT speaker connects + PA sink resolves, but
`server_connected = False` → operator-guidance card "lost bridge
transport" everywhere on test-VM 105.

Root cause: B5 of the src-layout migration renamed `bt_types.py`
(repo root) to `sendspin_bridge/bridge/types.py`. The subprocess is
launched with `cwd = os.path.dirname(__file__)` = bridge/ directory,
and Python's `-m module` invocation prepends cwd to sys.path[0].
When the daemon's `import functools` chain reaches `from types
import GenericAlias`, Python finds my `bridge/types.py` BEFORE
stdlib `types`, leaving the stdlib half-loaded and the import fails.

Fix: rename `bridge/types.py` back to `bridge/bt_types.py`. `types`
is a stdlib module name and shouldn't be reused by any package
whose modules might be added to sys.path[0] (which `python -m`
does at subprocess spawn time). Three call-site imports updated:
- demo/bt_manager.py
- src/sendspin_bridge/bluetooth/audio.py
- src/sendspin_bridge/bluetooth/manager.py

Verified: pytest 2205/2205 green.

Bumps VERSION → 2.66.0-rc.4 to re-trigger release.yml so a fresh
image with the fix gets published.
