+++ VERSION
@@ -1,1 +1,1 @@
-2.66.0-rc.3
+2.66.0-rc.4
