+++ demo/bt_manager.py
@@ -12,7 +12,7 @@ from demo.fixtures import DEMO_DEVICE_STATUS, get_demo_adapter
 if TYPE_CHECKING:
     from collections.abc import Callable
 
-    from sendspin_bridge.bridge.types import BluetoothManagerHost
+    from sendspin_bridge.bridge.bt_types import BluetoothManagerHost
 
 logger = logging.getLogger(__name__)
