+++ src/sendspin_bridge/bridge/bt_types.py
